+++ app/src/main/java/com/example/composerecipeapp/ui/user_interest/UserInterest.kt
@@ -13,8 +13,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavGraphBuilder
 import com.example.composerecipeapp.R
 import com.example.composerecipeapp.platform.navigation.navigator.AppMainNavigation
+import com.example.composerecipeapp.platform.navigation.navigator.NavComposable
 import com.example.composerecipeapp.platform.navigation.screens.MainViewIntent
 import com.example.composerecipeapp.platform.navigation.screens.UserInterestIntent
 import com.example.composerecipeapp.ui.provider.ParentNavHostController
@@ -23,6 +25,13 @@ import com.example.composerecipeapp.ui.theme.primaryColorDark
 import com.example.composerecipeapp.ui.views.CuisineList
 import com.example.composerecipeapp.util.observeState
 import com.example.composerecipeapp.viewmodel.user_interest.*
+
+@OptIn(ExperimentalFoundationApi::class)
+fun NavGraphBuilder.UserInterestScreen(){
+    NavComposable(UserInterestIntent()) {
+        UserInterest()
+    }
+}
 
 @ExperimentalFoundationApi
 @Composable
